utils: merge several jsutils into one
